requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.1.0+cu118
torchvision==0.16.0
datasets==2.14.5
huggingface-hub==0.17.3
lightning==2.0.9.post0
lightning-cloud==0.5.39
lightning-utilities==0.9.0
numpy==1.26.4
pandas==2.0.3
pytorch-lightning==2.0.9.post0
tokenizers==0.14.1
torchmetrics==1.2.0
tqdm==4.66.1
transformers==4.34.0
hydra-core==1.3.2
wandb==0.19.11
